// standmate/lib/types/generated-api.ts
export interface components {
    schemas: {
// ...
        ProjectCreate: {
            /** Team Id */
            team_id: number;
            properties: components["schemas"]["ProjectProperties"];
            /** Lead Id */
            lead_id?: number | null;
        };
        /** ProjectProperties */
        ProjectProperties: {
            /** Project Title */
            project_title: string;
            /** Short Summary */
            short_summary?: string | null;
            /**
             * Status
             * @default Backlog
             */
            status: string;
            /**
             * Priority
             * @default None
             */
            priority: string;
            /** Lead */
            lead?: string | number | null;
            /**
             * Members
             * @default []
             */
            members: (string | number)[];
            /** Start Date */
            start_date?: string | null;
            /** Target Date */
            target_date?: string | null;
            /**
             * Labels
             * @default []
             */
            labels: (string | number)[];
            /**
             * Dependencies
             * @default []
             */
            dependencies: (string | number)[];
            /** Description */
            description?: string | null;
            /**
             * Milestones
             * @default []
             */
            milestones: {
                [key: string]: unknown;
            } | unknown[];
        } & {
            [key: string]: unknown;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}